rename: flavor fix
