--- .github/ISSUE_TEMPLATE/🐞-버그-리포트-템플릿.md
@@ -1,25 +1,0 @@
----
-name: "\U0001F41E 버그 리포트 템플릿"
-about: 프로젝트에서 발생하는 버그에 대해 명세합니다.
-title: "[Bug] Issue title"
-labels: "\U0001F41E BugFix"
-assignees: ""
----
-
-## 💬 버그 설명
-
-문제가 생긴 상황을 간단히 설명해주세요.
-
-## 🔁 재현 방법
-
-문제가 어떻게 발생했는지 순서대로 작성해주세요.
-
-## ⚙️ 기대 동작
-
-정상적으로 어떤 동작을 기대했는지 설명해주세요.
-
-## 📸 스크린샷 (선택사항)
-
-## 📄 추가 정보
-
-기타 참고할만한 정보나 자료가 있다면 자유롭게 작성해주세요.
