+++ src/components/flavor/flavor.tsx
