+++ src/components/index.ts
@@ -2,7 +2,7 @@ export { default as Gnb } from "./gnb/gnb";
 export { default as Header } from "./header/header";
 export { default as DropdownMenu } from "./dropdown-menu/DropdownMenu";
 export { default as Chip } from "./chip/Chip";
-export { default as Flavor } from "./flavor/Flavor";
+export { default as Flavor } from "./flavor/flavor";
 export { default as TextInput } from "./text-input/text-input";
 export { default as SelectType } from "./select-type/select-type";
 export { default as Icon } from "./icon/Icon";
